Backend minesweeper init code
Mostly the backend logic for creating a minesweeper game. It contains a 10x10 board with a set number of bombs that are scrambled in each row of the board (not the best way of randomizing a board but works for testing). The left click is implemented but no right click (left click is clearing things that are not bombs while right click is declaring where a bomb is located but not trying to clear it).