Fixed count bomb issue and added a flag
in some situations, the count_bombs command would count too few bombs. This code was completely changed as the easiest way to count bombs is to just assume each square the person clicked at has 8 squares around it and check if it is possible to reach that square before checking if it is a bomb or not. The flag is the right click in a minesweeper game where you can flag where you think a mine is located. If the right click is pressed on a flag, it changes to a question mark. I don't know if you can left click a question mark or not so I left it as you cannot left click the question mark.